+++ lib/hooks/useClientAppOrigin.ts
@@ -1,0 +1,15 @@
+"use client";
+
+import { useSyncExternalStore } from "react";
+
+function subscribe() {
+  return () => {};
+}
+
+export function useClientAppOrigin() {
+  return useSyncExternalStore(
+    subscribe,
+    () => window.location.origin,
+    () => ""
+  );
+}